[SPARK-40216][SQL] Extract common `ParquetUtils.prepareWrite` method  to deduplicate code in `ParquetFileFormat` and `ParquetWrite`
### What changes were proposed in this pull request?
This pr is a refactor work, the main change is extract `ParquetUtils.prepareWrite` method to eliminate duplicate code in `ParquetFileFormat` and `ParquetWrite`.

### Why are the changes needed?
Eliminate duplicate code

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Pass GitHub Actions

Closes #37654 from LuciferYang/parquet-prepareWrite.

Authored-by: yangjie01 <yangjie01@baidu.com>
